+++ src/Lmc.ServiceIdentification/SimpleTypes.fs
@@ -3,14 +3,20 @@ namespace Lmc.ServiceIdentification
 [<RequireQualifiedAccess>]
 module Domain =
     let value (Domain domain) = domain
+    let map = SimpleType.map value Domain
+    let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module Context =
     let value (Context context) = context
+    let map = SimpleType.map value Context
+    let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module Purpose =
     let value (Purpose purpose) = purpose
+    let map = SimpleType.map value Purpose
+    let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module PurposePattern =
@@ -21,6 +27,8 @@ module PurposePattern =
 [<RequireQualifiedAccess>]
 module Version =
     let value (Version version) = version
+    let map = SimpleType.map value Version
+    let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module VersionPattern =
@@ -31,6 +39,8 @@ module VersionPattern =
 [<RequireQualifiedAccess>]
 module Zone =
     let value (Zone zone) = zone
+    let map = SimpleType.map value Zone
+    let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module ZonePattern =
@@ -41,6 +51,8 @@ module ZonePattern =
 [<RequireQualifiedAccess>]
 module Bucket =
     let value (Bucket bucket) = bucket
+    let map = SimpleType.map value Bucket
+    let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module BucketPattern =