+++ src/Lmc.ServiceIdentification/Utils.fs
@@ -12,3 +12,9 @@ module internal String =
         | null -> Some IsEmpty
         | empty when empty |> String.IsNullOrWhiteSpace -> Some IsEmpty
         | _ -> None
+
+    let toLower (string: string) = string.ToLower()
+
+[<RequireQualifiedAccess>]
+module internal SimpleType =
+    let map (fromType: 'SimpleType -> 'a) (toType: 'b -> 'SimpleType) (f: 'a -> 'b) = fromType >> f >> toType